requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==18.2.0
chardet==3.0.4
discord.py==1.3.3
disputils==0.1.2
idna==2.8
multidict==4.5.2
peewee==3.8.2
websockets==6.0
yarl==1.3.0
